+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io.kotest.engine/spec/materializeAndOrderRootTests.kt
@@ -1,6 +1,8 @@
-package io.kotest.core.spec
+package io.kotest.engine.spec
 
 import io.kotest.core.config.configuration
+import io.kotest.core.spec.RootTest
+import io.kotest.core.spec.Spec
 import io.kotest.core.test.TestCaseOrder
 
 /**
