+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io.kotest.engine/test/TestCaseExecutor.kt
@@ -1,6 +1,5 @@
 package io.kotest.engine.test
 
-import io.kotest.core.TimeoutExecutionContext
 import io.kotest.core.extensions.TestCaseExtension
 import io.kotest.core.extensions.resolvedTestCaseExtensions
 import io.kotest.core.internal.isEnabled
