+++ kotest-framework/kotest-framework-engine/src/commonTest/kotlin/com/sksamuel/kotest/engine/TestCaseOrderTest.kt
@@ -3,7 +3,7 @@ package com.sksamuel.kotest.engine
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.core.test.TestCaseOrder
-import io.kotest.core.spec.materializeAndOrderRootTests
+import io.kotest.engine.spec.materializeAndOrderRootTests
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
