+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/tags/TagsAnnotationInheritenceTest.kt
@@ -6,7 +6,7 @@ import io.kotest.core.config.configuration
 import io.kotest.core.extensions.TagExtension
 import io.kotest.core.internal.isEnabledInternal
 import io.kotest.core.spec.Isolate
-import io.kotest.core.spec.materializeAndOrderRootTests
+import io.kotest.engine.spec.materializeAndOrderRootTests
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
